Risky: Move tests into mod.rs
Usual Rust practice is to put tests in the same file they are testing.
I don't see any magic in the texttests that would require tests and code
to be in separate files, but I am not entirely sure, so I added the
risky tag.
