+++ rust/src/gildedrose/mod.rs
@@ -78,4 +78,15 @@ impl GildedRose {
 }
 
 #[cfg(test)]
-mod test;
+mod tests {
+    use super::{GildedRose, Item};
+
+    #[test]
+    pub fn foo() {
+        let items = vec![Item::new(String::from("foo"), 0, 0)];
+        let mut rose = GildedRose::new(items);
+        rose.update_quality();
+
+        assert_eq!("fixme", rose.items[0].name);
+    }
+}
